Reorganize files into folders
